Subversion project dptdb revision 3731 2017-02-13 08:43:30
Update dptdb to match pyDPT, including extend_dptapi_python.

pyDPT is the wrong place to extend the Python interface because pyDPT is
in it's state at the removal of www.dptoolkit.com and is marked Inactive
on PyPI.  These changes in dptdb are an attempt to avoid writing a lot
of DPT versions of CQL modules in ChessTab by extending APIRecordList
with & ^ | &= ^= and |= operators.

This is the first version that completes the build, but the C++ added
via Swig may not be correct.