Subversion project dptdb revision 4766 2021-10-13 08:28:28
Remove code identifying path to Python executable for Microsoft Windows.

Msys2 environment provides Python so use it.

The dptdb.dptapi module is importable by the Msys2-mingw32 Python which
built it.  This is Python 3.9 so it seems dptdb can be used in Microsoft
Windows at Python 3.8 and later provided the Msys2 environment, not
Msys, does the build.