Remove makefiles to compile DPT sources in VS 2017 x64 environment.
Adjust comments to be more explicit about x86 (32 bit) and x64 (64 bit)
versions.

SWIG wrappers for a Python interface in the x64 environment were built
and the wrapper module can be imported.  However runtime errors occur
almost immediately on use so the two makefiles involved in achieving
the x64 compliation are removed.