Add makefiles to build SWIG wrappers in VS 2017 x64 environment.
Previous two commits do not state situation with x64 crashes correctly.

There are runtime crashes, but a few actions succeed and the database
often remains sound for access by the x86 version of the wrappers.

In particular single step deferred update seems fine provided the update
does not need to go beyond the first segment (about 65000 records); and
a few database traversals succeed even on databases, populated by the x86
version, with a few million records.

The ChessTab project is used to try all this.

A mess, but not complete absence of hope for x64 dptdb.