Add populate database with 200000 records tests to package-template.
Index values are chosen so each references only one record on database.

This means sets of records referenced by an index value are not
represented as lists or bitmaps.

It was known the test runs would succeed on x86 environment; but needed
to be seen on x64 environment to eliminate some cases as causes of
repeatable crashes on x64.