Add 'cout' statements to trace 'Bit subscript ...' problem on x64.
The intention was to print argument values to locate where and what
arguments got set wrong.

What actually happened was adding the 'cout' statements caused the
critical test, run_test_traverse_create_3_delete_1, to complete
successfully.

This commit keeps the one, possibly two, 'cout' statements responsible
for the effect.  A manual edit must supply the '#include <iostream>'
statement needed because the pattern which restricts the edit to the
relevant code file is multi-line and cannot be handled by 'sed'.

The effect is seen in both Visual Studio and Msys2 x64 versions.

The next commit will reverse this change.