Built wheels renamed to correct python-, abi- and package- tags.
The Msys2 and Visual Studio 32 and 64 bit toolchains all name the
built wheels as <name>-<version>-py3-none-any.whl but each is tied
to the build environment so the relevant wheel names are:

<name>-<version>-cpXY-cpXY-mingw_i686.whl
<name>-<version>-cpXY-cpXY-mingw_x86_64.whl
<name>-<version>-cpXY-cpXY-win32.whl
<name>-<version>-cpXY-cpXY-win_amd64.whl

where X is Python major version and Y is Python minor version
such as '310'.

The Msys2 64 bit name was slightly hard to track down because it's
Python 3.10 is missing distutils: for distutils.util.get_platform().

PEP 632 suggests the platform module as source but nothing relevant
was found.

stackoverflow.com/questions/716648875/... suggests
sysconfig.get_platform() which seems to be the answer.

The Visual Studio 64 bit name is given as win-amd64 which cannot be
correct as the platform-tag because of the '-' rather than '_'.