Allow wheel version to vary for different package tag values.
It was, and still is, reasonable to give the 32 bit wheels a version
like '1.0.0'.  The 64 bit wheels cannot have such versions: for
these versions like '1.0.0.dev0' must be used at present if these
wheels are published at all.

PythonSed is installed from PyPI to enable this for win_amd64 and
win32 package tags built in Visual Studio environments.